+++ projects/ion/src/lib/ion.module.ts
@@ -30,14 +30,14 @@ import { SimpleMenuComponent } from './simple-menu/simple-menu.component';
 import { SmartTableComponent } from './smart-table/smart-table.component';
 import { TableComponent } from './table/table.component';
 import { IonModalComponent } from './modal/component/modal.component';
-import { DatePickerComponent } from './date-picker/picker/date-picker/date-picker.component';
+import { DatePickerComponent } from './picker/date-picker/calendar/date-picker.component';
 import { PopConfirmDirective } from './popconfirm/popconfirm.directive';
 import { PopConfirmComponent } from './popconfirm/popconfirm.component';
 import { InputComponent } from './input/input.component';
 import { TooltipDirective } from './tooltip/tooltip.directive';
 import { TooltipComponent } from './tooltip/tooltip.component';
-import { DatepickerComponent } from './date-picker/picker/datepicker.component';
-import { ControlPickerComponent } from './date-picker/control-picker/control-picker.component';
+import { ControlPickerComponent } from './picker/control-picker/control-picker.component';
+import { DatepickerComponent } from './picker/date-picker/datepicker.component';
 
 @NgModule({
   declarations: [
