+++ projects/ion/src/lib/picker/control-picker/control-picker.component.scss
