+++ projects/ion/src/lib/picker/date-picker/datepicker.component.html
@@ -18,7 +18,6 @@
   </div>
   <div class="content">
     <date-picker
-      *ngIf="pickerMode === 'datepicker'"
       [currentDate]="currentDate"
       [calendarControlAction]="calendarControlAction"
       [goToMonthInCalendar]="goToMonth"
