+++ projects/ion/src/lib/picker/date-picker/datepicker.component.scss
